Rename src to server
