+++ index.js
@@ -1,5 +1,6 @@
-import makeStore from './src/store';
-import startServer from './src/server';
+import makeStore from './server/store';
+import startServer from './server/server';
 
 export const store = makeStore();
+
 startServer(store);
