+++ test/store_spec.js
@@ -1,7 +1,7 @@
 import { Map, fromJS } from 'immutable';
 import { expect } from 'chai';
 
-import makeStore from '../src/store';
+import makeStore from '../server/store';
 
 describe('store', () => {
   it('is a Redux store configured with the correct reducer', () => {
